+++ src/Helpers/FormBuilder/FormTextarea.php
@@ -2,12 +2,13 @@
 
 namespace Larrock\Core\Helpers\FormBuilder;
 
-use View;
-
 class FormTextarea extends FBElement
 {
     /** @var null|bool */
     public $typo;
+
+    /** @var string Имя шаблона FormBuilder для отрисовки поля */
+    public $FBTemplate = 'larrock::admin.formbuilder.textarea.editor';
 
     /**
      * @return $this
@@ -25,19 +26,5 @@ class FormTextarea extends FBElement
     {
         $this->cssClass .= ' not-editor';
         return $this;
-    }
-
-    /**
-     * @param $row_settings
-     * @param $data
-     * @return mixed
-     */
-    public function render($row_settings, $data)
-    {
-        if( !isset($data->{$row_settings->name}) && $row_settings->default){
-            $data->{$row_settings->name} = $row_settings->default;
-        }
-        return View::make('larrock::admin.formbuilder.textarea.editor', ['row_key' => $row_settings->name,
-            'row_settings' => $row_settings, 'data' => $data])->render();
     }
 }